fix(packaging): close installed module dependencies
